requirements.txt
Django==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-axes==6.1.1
argon2-cffi==23.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
django-security==0.12.0
django-csp==3.7
python-decouple==3.8
Pillow==11.0.0
psycopg2-binary